feat: fix critical issues + add x402 payment support
Fixes from audit:
- main.rs: add TangleProducer + TangleConsumer (on-chain jobs were dead code)
- billing.rs: verify recovered signer address against expected spending key
  (was an auth bypass — any valid ECDSA signature passed)
- lib.rs: fix BackgroundService readiness signal

New features:
- x402 payment middleware on HTTP server: returns 402 with payment
  instructions, validates SpendAuth from X-Payment-Signature header
- Dual-path architecture: on-chain jobs (TangleProducer) + x402 HTTP
  requests feed the same Router and inference handler
- Single model per instance, model specified in service request config

24 contract tests pass, cargo check clean.

+++ contracts/src/InferenceBSM.sol
@@ -343,14 +343,14 @@ contract InferenceBSM is IBlueprintServiceManager {
     // ═══════════════════════════════════════════════════════════════════════
 
     /// @notice Validate an inference job submission
-    /// @dev inputs = abi.encode(string prompt, uint32 maxTokens, uint32 temperatureE4)
+    /// @dev inputs = abi.encode(string prompt, uint32 maxTokens, uint64 temperature)
     function onJobCall(
         uint64 serviceId,
         uint8,
         uint64 jobCallId,
         bytes calldata inputs
     ) external payable onlyFromTangle {
-        (, uint32 maxTokens,) = abi.decode(inputs, (string, uint32, uint32));
+        (, uint32 maxTokens,) = abi.decode(inputs, (string, uint32, uint64));
 
         emit InferenceJobSubmitted(serviceId, jobCallId, maxTokens);
     }
